added version
added version area. I used x's as I don't know what version you want.

It's for the format of:

1.0.0
1.7.10
0.0.1

You increase the small number by 1 for minor updates, the middle one for realitivly major, and the outer one for game-breaking (so 0 is typically pre-release (from early alpha to the end of beta), 1 is realease. 2 is typically a major re-write, but I increase it if I ever change multiple ID's or such.

Basically, from left to right:

almost guarenteed to be incompatible -> might be fine -> no problem

of course, it depends on the circumstances, I just guess each time.