requre
made it require that framework as you have features that require it, and they are active always.

It's possible to make features only active if certain mods are enabled, but honestly if you are using genes, there's nothing wrong with a framework! (And you can't iterate it easily, so you couldn't make something require both this framework and something else. There are ways around it, but they are, uh, finicky, at best)